requirements.txt
streamlit
torch
torchvision
pillow
opencv-python
numpy
SpeechRecognition
ffmpeg-python
